+++ src/http/httpd_cgi.c
@@ -96,7 +96,6 @@ static int hin_cgi_headers_read_callback (hin_buffer_t * buffer) {
   string_t line, orig=*source, param1, param2;
   http->status = 200;
   off_t sz = 0;
-  int cache = 0;
   uint32_t disable = 0;
   while (1) {
     if (find_line (source, &line) == 0) { return 0; }
@@ -112,12 +111,7 @@ static int hin_cgi_headers_read_callback (hin_buffer_t * buffer) {
       disable |= HIN_HTTP_CHUNKED;
     } else if (match_string (&line, "Cache%-Control:") > 0) {
       disable |= HIN_HTTP_CACHE;
-      int httpd_parse_cache_str (string_t * orig, hin_cache_data_t * cache);
-      hin_cache_data_t data;
-      memset (&data, 0, sizeof (data));
-      httpd_parse_cache_str (&line, &data);
-      http->cache = data.max_age;
-      cache = 1;
+      httpd_parse_cache_str (line.ptr, line.len, &http->cache_flags, &http->cache);
     } else if (matchi_string_equal (&line, "Date: .*") > 0) {
       disable |= HIN_HTTP_DATE;
     }
@@ -139,21 +133,21 @@ static int hin_cgi_headers_read_callback (hin_buffer_t * buffer) {
   pipe->parent1 = worker;
   pipe->finish_callback = hin_pipe_cgi_server_finish_callback;
 
-  if (cache && http->method != HIN_HTTP_POST && http->status == 200) {
+  if ((http->cache_flags & HIN_CACHE_PUBLIC) && http->method != HIN_HTTP_POST && http->status == 200) {
     // cache check is somewhere else
     int hin_cache_save (void * store, hin_pipe_t * pipe);
     int n = hin_cache_save (NULL, pipe);
     if (n == 0) {
       // pipe already started ?
       free (pipe);
-      return 0;
+      return -1;
     } else if (n > 0) {
       if (len > 0) {
         hin_buffer_t * buf1 = hin_buffer_create_from_data (pipe, source->ptr, len);
         hin_pipe_append (pipe, buf1);
       }
       hin_pipe_start (pipe);
-      return 0;
+      return -1;
     }
   }